chore(kreivo-runtime): refactor runtime_benchmarks apis
